Enhance secret configuration
Enhance configuration to support secret configuration values.
Secret configuration values are held in a secrets file.

In the "normal" configuration the secret configuration can be referenced
by the value "!secret <secret_key>".

The secret can only be accessed by config.secret("<key>") which follows
the indirection if there is any, otherwise it returns the current value.
If the secret value is accessed by  config.<key> and there is a secret
redirection "!secret <secret_key>" will be returned, thus hiding the secret.

Signed-off-by: Bobby Noelte <b0661n0e17e@gmail.com>